Add: files modified per mentor review; organized files into corresponding folders
