+++ public/data_yoojin/feedData.json
@@ -2,8 +2,8 @@
   {
     "id": 1,
     "profileImage": "https://scontent-ssn1-1.cdninstagram.com/v/t51.2885-15/e35/s320x320/33399786_217772972338995_4926975362853765120_n.jpg?_nc_ht=scontent-ssn1-1.cdninstagram.com&_nc_cat=103&_nc_ohc=535PyevtmuEAX_skI05&tp=1&oh=10eb37ee409785a9f97673bc3d2d72a6&oe=60422E85",
-    "proAlt": "profeed1",
-    "mainAlt": "mainfeed1",
+    "profileAlt": "profeed1",
+    "mainFeedAlt": "mainfeed1",
     "userName": "hi_dobie",
     "mainImage": "https://scontent-ssn1-1.cdninstagram.com/v/t51.2885-15/sh0.08/e35/s640x640/132198745_726606471614956_2757828674207918050_n.jpg?_nc_ht=scontent-ssn1-1.cdninstagram.com&_nc_cat=104&_nc_ohc=9te1DMY9QCkAX8HURBZ&tp=1&oh=a37ce0131a7a1ecbdb21801eee206150&oe=60406800",
     "content": "Dobie's biggest smile ^_^",
@@ -12,8 +12,8 @@
   {
     "id": 2,
     "profileImage": "https://scontent-ssn1-1.cdninstagram.com/v/t51.2885-15/e35/s320x320/130573006_1303926446620413_358204558991372102_n.jpg?_nc_ht=scontent-ssn1-1.cdninstagram.com&_nc_cat=109&_nc_ohc=tApWJAKa6ucAX9aUVF8&tp=1&oh=3b00cb00e8692f16b9f20c8139ae054a&oe=6041FFEC",
-    "proAlt": "profeed2",
-    "mainAlt": "mainfeed2",
+    "profileAlt": "profeed2",
+    "mainFeedAlt": "mainfeed2",
     "userName": "__rylieinNYC",
     "mainImage": "https://scontent-ssn1-1.cdninstagram.com/v/t51.2885-15/e35/60170269_671034343326829_2030083148076457786_n.jpg?_nc_ht=scontent-ssn1-1.cdninstagram.com&_nc_cat=101&_nc_ohc=a44V1NGDI7kAX8Hdfr3&tp=1&oh=5e20015b93729a1360b9c5061baccd3f&oe=60410E0B",
     "content": "Hey, look at all the buildings!",
@@ -23,8 +23,8 @@
   {
     "id": 3,
     "profileImage": "https://scontent-ssn1-1.cdninstagram.com/v/t51.2885-15/e35/s320x320/18444515_127938254435690_4739775029959458816_n.jpg?_nc_ht=scontent-ssn1-1.cdninstagram.com&_nc_cat=102&_nc_ohc=8Xu9T6XWsEsAX8BZbwe&tp=1&oh=82a0bc8821a146f493d5033b95c88f64&oe=60438683",
-    "proAlt": "profeed3",
-    "mainAlt": "mainfeed3",
+    "profileAlt": "profeed3",
+    "mainFeedAlt": "mainfeed3",
     "userName": "suukju____",
     "mainImage": "https://scontent-ssn1-1.cdninstagram.com/v/t51.2885-15/e35/c0.135.1080.1080a/s480x480/27576520_166039444045089_2397414301829693440_n.jpg?_nc_ht=scontent-ssn1-1.cdninstagram.com&_nc_cat=109&_nc_ohc=CGYVNrQmZHsAX-eKkU6&tp=1&oh=3ceaf11c2390b68b5287df21cd52dc5c&oe=6040B931",
     "content": "lost in wind",
