+++ src/Pages/yoojinsuh/Component/Comment/Comment.js
@@ -5,27 +5,26 @@ class Comment extends Component {
   constructor() {
     super();
     this.state = {
-      commentList: [],
       heartFilled: false,
     };
   }
 
   fillHeart = () => {
     this.setState({
-      heartFilled: !this.state.heartFilled ? true : false,
+      heartFilled: !this.state.heartFilled,
     });
   };
 
   render() {
     const { id, comment, deleteComment } = this.props;
     return (
-      <li className="Comment" key={id}>
+      <li className="Comment">
         <div>
           <span className="userAccount">hi_yoojins</span>
           <span> {comment}</span>
         </div>
         <div className="heartAndTrash">
-          <i className="far fa-trash-alt" id={id} onClick={deleteComment} />
+          <i className="far fa-trash-alt" onClick={() => deleteComment(id)} />
           <i
             className="far fa-heart"
             id={this.state.heartFilled ? "heartFilled" : "heartEmpty"}
