+++ src/Pages/yoojinsuh/Component/Comment/Comment.scss
@@ -2,8 +2,7 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  margin-top: 3px;
-  margin-bottom: 5px;
+  margin: 3px auto 5px;
 
   .userAccount {
     font-weight: bold;
