+++ src/Pages/yoojinsuh/Component/Feed/Feed.scss
@@ -1,0 +1,161 @@
+.Feed {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  margin-top: 25px;
+  border: 1px solid #dbdbdb;
+  border-radius: 2px;
+  width: 600px;
+
+  header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    height: 60px;
+
+    .accountContainer {
+      display: flex;
+      justify-content: flex-start;
+      align-items: center;
+
+      h1 {
+        width: 476px;
+        padding-left: 13px;
+        color: black;
+        text-align: left;
+        font-weight: bold;
+      }
+
+      img {
+        width: 45px;
+        margin-left: 15px;
+        border-radius: 30px;
+        padding: 2px;
+        border: 2px solid #c62e61;
+      }
+    }
+
+    i {
+      margin-right: 20px;
+      color: #595757;
+    }
+  }
+
+  .mainImage {
+    width: 100%;
+    height: 600px;
+  }
+
+  footer {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    height: 40px;
+    margin: 0 20px;
+
+    i {
+      width: 20px;
+      font-size: 25px;
+    }
+
+    .footerNav {
+      display: flex;
+      justify-content: space-between;
+      width: 100px;
+
+      i {
+        margin-right: 20px;
+      }
+
+      #heartEmpty:hover {
+        font-weight: bold;
+        color: red;
+      }
+
+      #heartFilled {
+        font-weight: bold;
+        color: red;
+      }
+
+      #heartEmpty {
+        color: black;
+      }
+    }
+
+    #bookmark:hover {
+      font-weight: bold;
+      color: black;
+    }
+  }
+
+  .likes {
+    margin-left: 20px;
+    font-weight: bold;
+  }
+
+  .contentContainer {
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    padding-top: 8px;
+    max-height: 25px;
+    margin-left: 20px;
+
+    .accountName {
+      margin: 0 10px 5px 0;
+      font-weight: bold;
+    }
+
+    .contentDetail {
+      margin-bottom: 5px;
+    }
+  }
+
+  .comments {
+    .commentsLog {
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+      margin: auto 20px;
+
+      .commentsContainer {
+        list-style-type: none;
+
+        p {
+          margin-top: 15px;
+          margin-bottom: 15px;
+          color: #959595;
+          font-size: 10px;
+        }
+      }
+    }
+
+    .addComment {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      height: 50px;
+      padding: 15px 20px;
+      border: none;
+
+      .inputComment {
+        width: 100%;
+        border: none;
+        outline: none;
+      }
+
+      .inputSubmit {
+        width: 30px;
+        color: #b0cfee;
+        border: none;
+      }
+
+      .darkBtn {
+        color: #1c7ed6;
+        width: 30px;
+        border: none;
+      }
+    }
+  }
+}
