+++ src/Pages/yoojinsuh/Component/Suggestion/Suggestion.scss
@@ -1,28 +1,30 @@
+$grey-bg-color: #fafafa;
+
 .suggestProfiles {
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
-  background-color: #fafafa;
+  background-color: $grey-bg-color;
 
   .profile {
     display: flex;
     align-items: center;
     justify-content: space-between;
     margin-bottom: 10px;
-    background-color: #fafafa;
+    background-color: $grey-bg-color;
   }
 
   .imageAndName {
     display: flex;
     align-items: center;
     justify-content: flex-start;
-    background-color: #fafafa;
+    background-color: $grey-bg-color;
 
     .profileContainer {
       display: flex;
       flex-direction: column;
       justify-content: flex-start;
-      background-color: #fafafa;
+      background-color: $grey-bg-color;
     }
 
     img {
@@ -32,21 +34,21 @@
     }
 
     .name {
-      background-color: #fafafa;
+      background-color: $grey-bg-color;
       font-weight: bold;
     }
 
     .detail {
       margin-top: 5px;
       color: rgb(145, 144, 144);
-      background-color: #fafafa;
+      background-color: $grey-bg-color;
       font-size: 12px;
     }
   }
 
   .follow {
     color: #0095f6;
-    background-color: #fafafa;
+    background-color: $grey-bg-color;
     font-size: 13px;
   }
