+++ src/Pages/yoojinsuh/Login/Login.scss
@@ -67,8 +67,7 @@
     justify-content: space-between;
     width: 100%;
     height: 30px;
-    margin-top: 0px;
-    margin-bottom: 16px;
+    margin: 0px 0px 16px 0px;
     padding-top: 5px;
 
     .left,
