+++ src/Pages/yoojinsuh/Main/Main.scss
@@ -1,166 +1,176 @@
 @import url("https://use.fontawesome.com/releases/v5.15.2/css/all.css");
 @import url("https://fonts.googleapis.com/css2?family=Lato:wght@300;400;700&family=Lobster+Two:ital@1&display=swap");
 
+$white-bg-color: #ffffff;
+
 * {
   box-sizing: border-box;
   margin: 0;
   padding: 0;
   color: black;
-  background-color: #ffffff;
+  background-color: $white-bg-color;
   font-size: 15px;
   font-weight: normal;
   font-family: "Lato", sans-serif;
 }
 
-.navbar {
-  position: fixed;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 45px;
-  width: 100%;
-  border-bottom: 1px solid #dbdbdb;
-
-  p {
-    margin-left: 20px;
-    font-size: 25px;
-    font-family: "Lobster Two", cursive;
-  }
-
-  p:hover {
-    transform: scale(1.1);
-    cursor: pointer;
-  }
-
-  .searchContainer {
-    position: relative;
-
-    input {
-      width: 230px;
-      height: 23px;
-      padding: 13px 0px 13px 0px;
-      color: #b6b4b4;
-      background-color: #fafafae5;
-      border: 1px solid #dbdbdb;
-      border-radius: 2px;
-      font-size: 14px;
-      text-align: center;
-    }
-
-    i {
-      position: absolute;
-      width: 10px;
-      margin-top: 7px;
-      margin-left: 75px;
-      color: #b6b4b4;
-    }
-  }
-
-  nav {
+.Main {
+  .navbar {
+    position: fixed;
     display: flex;
-    flex-direction: row;
     justify-content: space-between;
-    width: 150px;
-    margin-right: 20px;
+    align-items: center;
+    height: 45px;
+    width: 100%;
+    border-bottom: 1px solid #dbdbdb;
 
-    i {
+    p {
+      margin-left: 20px;
       font-size: 25px;
+      font-family: "Lobster Two", cursive;
     }
 
-    i:hover {
+    p:hover {
       transform: scale(1.1);
       cursor: pointer;
     }
-  }
-}
 
-main {
-  display: flex;
-  justify-content: center;
-  height: 1500px;
-  background-color: #fafafa;
+    .searchContainer {
+      position: relative;
 
-  .main-left {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    background-color: #fafafa;
-  }
-
-  .main-right {
-    height: 100%;
-    width: 400px;
-    margin-top: 30px;
-    margin-right: 20px;
-    margin-left: 30px;
-    background-color: #fafafa;
-
-    .myProfile {
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-      justify-content: flex-start;
-      width: 100%;
-      height: 120px;
-      margin-top: 30px;
-      background-color: #fafafa;
-
-      img {
-        width: 70px;
-        border-radius: 70px;
+      input {
+        width: 230px;
+        height: 23px;
+        padding: 13px 0px 13px 0px;
+        color: #b6b4b4;
+        background-color: #fafafae5;
+        border: 1px solid #dbdbdb;
+        border-radius: 2px;
+        font-size: 14px;
+        text-align: center;
       }
 
-      .accountInfo {
-        margin-left: 15px;
-        background-color: #fafafa;
-
-        h1,
-        h2 {
-          background-color: #fafafa;
-          text-align: left;
-        }
-
-        h1 {
-          font-weight: bold;
-        }
-
-        h2 {
-          margin-top: 6px;
-          color: rgb(145, 144, 144);
-        }
+      i {
+        position: absolute;
+        width: 10px;
+        margin: 7px 0px 0px 75px;
+        color: #b6b4b4;
       }
     }
 
-    .suggestions {
+    nav {
       display: flex;
-      flex-direction: column;
-      justify-content: flex-start;
-      width: 100%;
-      height: 260px;
-      margin-top: 20px;
+      flex-direction: row;
+      justify-content: space-between;
+      width: 150px;
+      margin-right: 20px;
+
+      i {
+        font-size: 25px;
+      }
+
+      i:hover {
+        transform: scale(1.1);
+        cursor: pointer;
+      }
+    }
+  }
+
+  main {
+    display: flex;
+    justify-content: center;
+    height: 1500px;
+    background-color: #fafafa;
+
+    .mainLeft {
+      .storyContainer {
+        overflow: scroll auto;
+        display: flex;
+        align-items: center;
+        justify-content: flex-start;
+        width: 598px;
+        height: 120px;
+        margin-top: 60px;
+        border: 1px solid #dbdbdb;
+        border-radius: 2px;
+      }
+    }
+
+    .mainRight {
+      height: 100%;
+      width: 400px;
+      margin-top: 30px;
+      margin-right: 20px;
+      margin-left: 30px;
       background-color: #fafafa;
 
-      .header {
+      .myProfile {
         display: flex;
         flex-direction: row;
-        justify-content: space-between;
         align-items: center;
+        justify-content: flex-start;
         width: 100%;
-        height: 30px;
-        margin-bottom: 10px;
+        height: 120px;
+        margin-top: 30px;
         background-color: #fafafa;
 
-        .suggestForYou {
-          color: rgb(145, 144, 144);
-          background-color: #fafafa;
-          font-weight: bold;
+        img {
+          width: 70px;
+          border-radius: 70px;
         }
 
-        .seeAll {
-          color: #262626;
+        .accountInfo {
+          margin-left: 15px;
           background-color: #fafafa;
-          font-weight: bold;
-          font-size: 13px;
+
+          h1,
+          h2 {
+            background-color: #fafafa;
+            text-align: left;
+          }
+
+          h1 {
+            font-weight: bold;
+          }
+
+          h2 {
+            margin-top: 6px;
+            color: rgb(145, 144, 144);
+          }
+        }
+      }
+
+      .suggestions {
+        display: flex;
+        flex-direction: column;
+        justify-content: flex-start;
+        width: 100%;
+        height: 260px;
+        margin-top: 20px;
+        background-color: #fafafa;
+
+        .header {
+          display: flex;
+          flex-direction: row;
+          justify-content: space-between;
+          align-items: center;
+          width: 100%;
+          height: 30px;
+          margin-bottom: 10px;
+          background-color: #fafafa;
+
+          .suggestForYou {
+            color: rgb(145, 144, 144);
+            background-color: #fafafa;
+            font-weight: bold;
+          }
+
+          .seeAll {
+            color: #262626;
+            background-color: #fafafa;
+            font-weight: bold;
+            font-size: 13px;
+          }
         }
       }
     }
